terminado

+++ pages/api/products/[id].js
@@ -4,8 +4,11 @@ import { db } from "../../../database";
 export default async function (req, res) {
    const {
       method,
+      cookies,
       query: { id },
    } = req;
+
+   const token = cookies.token;
 
    await db.connect();
 
@@ -22,6 +25,10 @@ export default async function (req, res) {
    }
 
    if (method === "PUT") {
+      if (!token || token !== process.env.TOKEN) {
+         return res.status(401).json("Not authenticated");
+      }
+
       const product = await Product.findByIdAndUpdate(id, req.body, {
          new: true,
       });
@@ -39,6 +46,10 @@ export default async function (req, res) {
    }
 
    if (method === "DELETE") {
+      if (!token || token !== process.env.TOKEN) {
+         return res.status(401).json("Not authenticated");
+      }
+
       const product = await Product.findByIdAndDelete(id);
 
       res.status(200).json("Product delete: ", product);
